+++ packages/plugin-ext/src/hosted/browser/worker/worker-main.ts
@@ -19,7 +19,7 @@ import { RPCProtocolImpl } from '../../../common/rpc-protocol';
 import { PluginManagerExtImpl } from '../../../plugin/plugin-manager';
 import { MAIN_RPC_CONTEXT, Plugin, emptyPlugin } from '../../../common/plugin-api-rpc';
 import { createAPIFactory } from '../../../plugin/plugin-context';
-import { getPluginId, PluginMetadata } from '../../../common/plugin-protocol';
+import { getPluginId, PluginMetadata, PluginPackage, PluginModel } from '../../../common/plugin-protocol';
 import * as theia from '@theia/plugin';
 import { PreferenceRegistryExtImpl } from '../../../plugin/preference-registry';
 import { ExtPluginApi } from '../../../common/plugin-ext-api-contribution';
@@ -58,6 +58,20 @@ const preferenceRegistryExt = new PreferenceRegistryExtImpl(rpc, workspaceExt);
 const debugExt = createDebugExtStub(rpc);
 const clipboardExt = new ClipboardExt(rpc);
 
+function stubPluginPackage(pluginModel: PluginModel): PluginPackage {
+    return {
+        name: pluginModel.name,
+        displayName: pluginModel.displayName,
+        engines: {
+            [pluginModel.engine.type]: pluginModel.engine.version
+        },
+        packagePath: pluginModel.packagePath,
+        publisher: pluginModel.publisher,
+        version: pluginModel.version,
+        description: ''
+    };
+}
+
 const pluginManager = new PluginManagerExtImpl({
     // tslint:disable-next-line:no-any
     loadPlugin(plugin: Plugin): any {
@@ -90,10 +104,10 @@ const pluginManager = new PluginManagerExtImpl({
                 }
                 const plugin: Plugin = {
                     pluginPath: pluginModel.entryPoint.frontend!,
-                    pluginFolder: plg.source.packagePath,
+                    pluginFolder: pluginModel.packagePath,
                     model: pluginModel,
                     lifecycle: pluginLifecycle,
-                    rawModel: plg.source
+                    rawModel: stubPluginPackage(pluginModel)
                 };
                 result.push(plugin);
                 const apiImpl = apiFactory(plugin);
@@ -102,10 +116,10 @@ const pluginManager = new PluginManagerExtImpl({
             } else {
                 foreign.push({
                     pluginPath: pluginModel.entryPoint.backend!,
-                    pluginFolder: plg.source.packagePath,
+                    pluginFolder: pluginModel.packagePath,
                     model: pluginModel,
                     lifecycle: pluginLifecycle,
-                    rawModel: plg.source
+                    rawModel: stubPluginPackage(pluginModel)
                 });
             }
         }